CLMFAP_Ver.2/pip_requirements.txt
absl-py==2.1.0
aiohappyeyeballs==2.3.5
aiohttp==3.10.3
aiosignal==1.3.1
antlr4-python3-runtime==4.8
apex==0.1
async-timeout==4.0.3
attrs==24.2.0
bitarray==2.9.2
bleach==6.2.0
bokeh==3.4.3
bytecode==0.16.0
certifi==2024.7.4
cffi==1.17.0
charset-normalizer==3.3.2
click==8.1.7
cloudpickle==3.1.0
cmake==3.30.2
colorama==0.4.6
colorcet==3.1.0
contourpy==1.3.0
cycler==0.12.1
Cython==3.0.11
dask==2024.8.0
datashader==0.16.3
einops==0.8.0
fairseq==0.12.2
filelock==3.15.4
fonttools==4.53.1
frozenlist==1.4.1
fsspec==2024.6.1
future==1.0.0
grpcio==1.65.4
holoviews==1.20.0
huggingface-hub==0.24.7
hydra-core==1.0.7
idna==3.7
imageio==2.36.0
importlib_metadata==8.2.0
importlib_resources==6.4.5
Jinja2==3.1.4
joblib==1.4.2
kiwisolver==1.4.7
lazy_loader==0.4
linkify-it-py==2.0.3
lit==18.1.8
littleutils==0.2.4
llvmlite==0.43.0
locket==1.0.0
lxml==5.3.0
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.8.0
mdit-py-plugins==0.4.2
mdurl==0.1.2
mpmath==1.3.0
multidict==6.0.5
multipledispatch==1.0.0
networkx==3.2.1
numba==0.60.0
numpy==1.26.3
nvidia-cublas-cu11==11.11.3.6
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu11==11.8.87
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu11==11.8.89
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu11==11.8.89
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu11==8.7.0.84
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu11==10.9.0.58
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu11==10.3.0.86
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu11==11.4.1.48
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu11==11.7.5.86
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu11==2.20.5
nvidia-nccl-cu12==2.18.1
nvidia-nvjitlink-cu12==12.6.20
nvidia-nvtx-cu11==11.8.86
nvidia-nvtx-cu12==12.1.105
ogb==1.3.6
omegaconf==2.0.6
outdated==0.2.2
packaging==24.1
pandas==2.2.3
panel==1.4.5
param==2.1.1
partd==1.4.2
pillow==10.2.0
pip==24.0
portalocker==2.10.1
protobuf==4.25.4
psutil==6.0.0
ptflops==0.7.3
pycparser==2.22
pyct==0.5.0
pynndescent==0.5.13
pyparsing==3.1.2
python-dateutil==2.9.0.post0
pytorch-fast-transformers==0.4.0
pytorch-lightning==1.1.5
pytz==2024.1
pyviz_comms==3.0.3
PyYAML==6.0.2
rdkit==2022.3.3
regex==2024.7.24
requests==2.32.3
sacrebleu==2.4.2
sacremoses==0.1.1
safetensors==0.4.4
scikit-image==0.24.0
scikit-learn==1.5.1
scipy==1.13.1
seaborn==0.13.2
setuptools==72.1.0
six==1.16.0
sympy==1.13.1
tabulate==0.9.0
tbb==2022.0.0
tcmlib==1.2.0
tensorboard==2.17.0
tensorboard-data-server==0.7.2
threadpoolctl==3.5.0
tifffile==2024.8.30
tokenizers==0.19.1
toolz==1.0.0
torch==2.3.0+cu118
torch_geometric==2.3.1
torch_scatter==2.1.1
torchaudio==2.3.0+cu118
torchsummary==1.5.1
torchvision==0.18.0+cu118
tornado==6.4.1
tqdm==4.66.5
transformers==4.44.2
triton==2.3.0
typing_extensions==4.12.2
tzdata==2024.2
uc-micro-py==1.0.3
umap-learn==0.5.1
urllib3==2.2.2
visualizer==0.0.1
webencodings==0.5.1
Werkzeug==3.0.3
wheel==0.43.0
xarray==2024.7.0
xyzservices==2024.9.0
yarl==1.9.4
zipp==3.19.2
